+++ Assets/Scripts/UDPToolkit/UDPClient.cs
@@ -157,11 +157,6 @@ namespace ubv
                         m_receivers[i].ReceivePacket(packet);
                     }
                 }
-
-                public void Disconnect()
-                {
-                    m_client.Close();
-                }
             }
         }
     }
